tls configuration moved to its own file and added mapstructure annotations
